registration form
functioning registration form, functioning input fields, password requirement with a min password req, ability to upload a profile picture & bio, as well as a working drop down menu - you will not be able to register unless you follow the correct password format & accept the terms and conditions. There is also a requirement for the age, it cannot be below 13 and above 120. I've also posted the css file that I used to style this form.